+++ apps/builder/src/widgetLibrary/TableWidget/utils.ts
@@ -1,4 +1,5 @@
 import { ColumnItemShape } from "@/widgetLibrary/TableWidget/interface"
+import { dayjsPro, isNumber } from "@illa-design/system"
 
 export const tansTableDataToColumns = (
   data: Record<any, any>[],
@@ -17,4 +18,31 @@ export const tansTableDataToColumns = (
     })
   }
   return columns
+}
+
+export const transDataForType = (data: ColumnItemShape) => {
+  const {
+    type = "text",
+    decimalPlaces = 0,
+    format = "YYYY-MM-DD",
+    header,
+  } = data
+  const cellValue = header
+  switch (type) {
+    default:
+      return cellValue
+    case "text":
+      return cellValue
+    case "number":
+      const formatVal = Number(cellValue)
+      return isNumber(formatVal) ? formatVal.toFixed(decimalPlaces) : "-"
+    case "percent":
+      const percentVal = Number(cellValue)
+      return isNumber(percentVal)
+        ? `${(percentVal * 100).toFixed(decimalPlaces)}%`
+        : "-"
+    case "date":
+      const dayVal = dayjsPro(cellValue).format(format)
+      return dayVal ? dayVal : "-"
+  }
 }
